requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.2
django-appconf==1.0.3
django-ckeditor==5.7.1
django-image-cropping==1.2.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-multiselectfield==0.1.8
django-schedule==0.5b0
django-social-share==1.3.2
django-static-precompiler==1.8.2
easy-thumbnails==2.6
geojson==2.4.1
gunicorn==19.9.0
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pilkit==2.0
Pillow==6.0.0
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pyowm==2.10.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
vobject==0.9.6.1
whitenoise==4.1.3
wrapt==1.11.1
